python/google_sheets/pyproject.toml
[project]
name = "google-sheets-tool"
version = "0.1.0"
description = "Google Sheets tool to fetch rows"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115.13",
    "optimizely-opal-opal-tools-sdk>=0.1.1.dev0",
    "httpx>=0.27.0",
]
